UI Work

+++ BotMonitor/Models/Instruction.cs
@@ -4,10 +4,10 @@ namespace BotMonitor.Models
 {
     public class Instruction
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
         
         [ForeignKey("Bot")]
-        public byte BotId { get; set; }
+        public int BotId { get; set; }
 
         public Bot Bot { get; set; }
